Выполнено задание 6-2
