+++ src/components/announcement-empty/announcement-empty.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+const АnnouncementEmpty = () => {
+  return (
+    <section className="cities__no-places">
+      <div className="cities__status-wrapper tabs__content">
+        <b className="cities__status">No places to stay available</b>
+        <p className="cities__status-description">We could not find any property available at the moment in Dusseldorf</p>
+      </div>
+    </section>
+  );
+};
+
+export default АnnouncementEmpty;
